chore: optimise planning constraints requests (#3740)
